Zenodo monitoring exporter layout
