--- assets/font.css
@@ -1,91 +1,0 @@
-/* Poppins */
-
-@font-face {
-    font-family: 'Poppins-SemiBoldItalic';
-    src: url('./Poppins/Poppins-SemiBoldItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Regular';
-    src: url('./Poppins/Poppins-Regular.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Black';
-    src: url('./Poppins/Poppins-Black.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-ExtraLight';
-    src: url('./Poppins/Poppins-ExtraLight.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-BlackItalic';
-    src: url('./Poppins/Poppins-BlackItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Light';
-    src: url('./Poppins/Poppins-Light.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Thin';
-    src: url('./Poppins/Poppins-Thin.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Medium';
-    src: url('./Poppins/Poppins-Medium.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-BoldItalic';
-    src: url('./Poppins/Poppins-BoldItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-LightItalic';
-    src: url('./Poppins/Poppins-LightItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-ExtraBold';
-    src: url('./Poppins/Poppins-ExtraBold.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Bold';
-    src: url('./Poppins/Poppins-Bold.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-ThinItalic';
-    src: url('./Poppins/Poppins-ThinItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-Italic';
-    src: url('./Poppins/Poppins-Italic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-ExtraLightItalic';
-    src: url('./Poppins/Poppins-ExtraLightItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-SemiBold';
-    src: url('./Poppins/Poppins-SemiBold.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-ExtraBoldItalic';
-    src: url('./Poppins/Poppins-ExtraBoldItalic.ttf') format('truetype');
-}
-
-@font-face {
-    font-family: 'Poppins-MediumItalic';
-    src: url('./Poppins/Poppins-MediumItalic.ttf') format('truetype');
-}
