--- assets/index.css
@@ -1,10 +1,0 @@
-/* Poppins */
-
-body {
-    font-family: 'Poppins-Regular';
-}
-
-h1 {
-    font-family: 'Poppins-Black';
-    text-align: center;
-}
